requirements.txt
requests
beautifulsoup4
python-dotenv
streamlit
cloudscraper==1.2.71
